requirements.txt
discord.py
mcipc